Filesource meta-table that lists all files
